Jour / Nuit, lumière des structures, temps en ms au lieu de secondes.

+++ core/main/src/com/pixurvival/core/contentPack/item/EdibleItem.java
@@ -14,7 +14,7 @@ public class EdibleItem extends Item {
 
 	private static final long serialVersionUID = 1L;
 
-	private double duration;
+	private long duration;
 	private List<Alteration> alterations = new ArrayList<>();
 
 }
